+++ CHANGELOG.md
@@ -2,6 +2,14 @@ Changelog
 ==============
 
 
+[v0.4](https://github.com/xtr4nge/FruityC2/releases/tag/v0.4)
+----------------
+- usemodule has been implemented (job)
+- Listener port issue has been fixed
+- Listener HTTPS option has been implemented
+- Commands list/description has been added
+
+
 [v0.3](https://github.com/xtr4nge/FruityC2/releases/tag/v0.3)
 ----------------
 - Agent background task has been implemented (powershell_job, shell_job)